chore: nodeHandler use context
