+++ nerdlets/observability-maps-nerdlet/components/custom-nodes/default.js
@@ -8,6 +8,7 @@ import {
 } from '../../lib/helper';
 import { LineChart, AreaChart, Billboard, PieChart, TableChart } from 'nr1';
 import { DataConsumer } from '../../context/data';
+import { buildNodeMetrics } from './node-utills';
 
 // const languageIcons = {
 //   java: "https://image.flaticon.com/icons/svg/226/226777.svg",
@@ -170,206 +171,111 @@ export default class CustomNode extends React.PureComponent {
     );
   }
 
-  render() {
-    const { node, mapData, nodeSize, closeCharts } = this.props;
-    const data = ((mapData || {}).nodeData || {})[node.id] || {};
+  renderIconGroup = (
+    userIcons,
+    data,
+    nodeId,
+    metrics,
+    icon,
+    iconOne,
+    colorTwo,
+    colorOne,
+    closeCharts
+  ) => {
+    const isOpen = this.state[`popup_${nodeId}`] || false;
+    const iconOuter = `circle ${iconOne}`;
 
-    // need logic to determine node health
-    const style = {
-      borderRadius: 0
-      // opacity: 0.7,
-      // padding: '2em',
-    };
-
-    const icon = data.icon || setEntityDesign(data.entityType).icon;
-    const metrics = [];
-
-    if (
-      data.hoverType === 'customNrql' &&
-      data.hoverData &&
-      data.hoverData.length > 0
-    ) {
-      data.hoverData.forEach(item => {
-        metrics.push({ value: item.value, name: item.name, unit: '' });
-      });
-    } else if (data.apmSummary) {
-      if (data.apmSummary.throughput)
-        metrics.push({
-          value: data.apmSummary.throughput,
-          unit: 'rpm',
-          name: 'throughput'
-        });
-      if (data.apmSummary.responseTimeAverage)
-        metrics.push({
-          value: data.apmSummary.responseTimeAverage,
-          unit: 'ms',
-          name: 'latency'
-        });
-      if (data.apmSummary.errorRate)
-        metrics.push({
-          value: data.apmSummary.errorRate,
-          unit: '%',
-          name: 'errors'
-        });
-    } else if (data.mobileSummary) {
-      if (data.mobileSummary.httpRequestRate)
-        metrics.push({
-          value: data.mobileSummary.httpRequestRate,
-          unit: 'rpm',
-          name: 'request rate'
-        });
-      if (data.mobileSummary.httpResponseTimeAverage)
-        metrics.push({
-          value: data.mobileSummary.httpResponseTimeAverage,
-          unit: 'ms',
-          name: 'latency'
-        });
-      if (data.mobileSummary.httpErrorRate)
-        metrics.push({
-          value: data.mobileSummary.httpErrorRate,
-          unit: '%',
-          name: 'errors'
-        });
-    } else if (data.externalSummary) {
-      if (data.externalSummary.throughput)
-        metrics.push({
-          value: data.externalSummary.throughput,
-          unit: 'rpm',
-          name: 'throughput'
-        });
-      if (data.externalSummary.responseTimeAverage)
-        metrics.push({
-          value: data.externalSummary.responseTimeAverage,
-          unit: 'ms',
-          name: 'latency'
-        });
-    } else if (data.browserSummary) {
-      // ajaxRequestThroughput: 49
-      // ajaxResponseTimeAverage: 0.156681
-      // jsErrorRate: 9.375
-      // pageLoadThroughput: 53.333333
-      // pageLoadTimeAverage: 1.292042
-      // pageLoadTimeMedian: 0.857
-      // spaResponseTimeAverage: 1.204
-      // spaResponseTimeMedian: 0.871
-      if (data.browserSummary.spaResponseTimeAverage)
-        metrics.push({
-          value: data.browserSummary.spaResponseTimeAverage,
-          unit: 'ms',
-          name: 'spa latency'
-        });
-      if (data.browserSummary.pageLoadTimeAverage)
-        metrics.push({
-          value: data.browserSummary.pageLoadTimeAverage,
-          unit: 'ms',
-          name: 'page latency'
-        });
-      if (data.browserSummary.pageLoadThroughput)
-        metrics.push({
-          value: data.browserSummary.pageLoadThroughput,
-          unit: 'rpm',
-          name: 'page load throughput'
-        });
-      if (data.browserSummary.jsErrorRate)
-        metrics.push({
-          value: data.browserSummary.jsErrorRate,
-          unit: '%',
-          name: 'js errors'
-        });
-    }
-
-    let { colorOne, colorTwo, iconOne } = setAlertDesign(
-      data.alertSeverity,
-      data.entityType
-    );
-    if (data.customAlert && data.customAlertData && data.customAlertData[0]) {
-      const customAlertDesign = setCustomAlertDesign(
-        data.customAlert,
-        data.customAlertData
-      );
-      colorOne = customAlertDesign.colorOne;
-      colorTwo = customAlertDesign.colorTwo;
-    }
-
-    const renderIconGroup = (
-      userIcons,
-      data,
-      nodeId,
-      metrics,
-      icon,
-      colorTwo,
-      colorOne,
-      closeCharts
-    ) => {
-      const isOpen = this.state[`popup_${nodeId}`] || false;
-      const iconOuter = `circle ${iconOne}`;
-
-      return (
-        <Icon.Group size="big">
-          {metrics && metrics.length > 0 ? (
-            <Popup
-              className="popup-custom"
-              trigger={
-                <Icon loading size="big" color={colorTwo} name={iconOuter} />
-              }
-              // on="click"
-              style={style}
-              inverted
-              hoverable
-              mouseLeaveDelay={3000}
-              content={this.renderContent(metrics)}
-            />
-          ) : (
-            <Icon loading size="big" color={colorTwo} name={iconOuter} />
-          )}
-
+    return (
+      <Icon.Group size="big">
+        {metrics && metrics.length > 0 ? (
           <Popup
-            // className="popup-custom"
-            trigger={this.renderIcon(
-              userIcons,
-              data,
-              nodeId,
-              isOpen,
-              icon,
-              colorOne,
-              colorTwo
-            )}
-            on="click"
-            style={style}
-            inverted
-            // onClose={() => this.setState({[`popup_${nodeId}`]:false})}
-            // onOpen={() => this.setState({[`popup_${nodeId}`]:true})}
-            open={
-              this.state[`popup_${nodeId}`] === true && closeCharts === false
+            className="popup-custom"
+            trigger={
+              <Icon loading size="big" color={colorTwo} name={iconOuter} />
             }
-            content={this.renderChart(data.mainChart)}
-            position="bottom center"
+            // on="click"
+            style={{ borderRadius: 0 }}
+            inverted
+            hoverable
+            mouseLeaveDelay={3000}
+            content={this.renderContent(metrics)}
           />
-        </Icon.Group>
-      );
-    };
+        ) : (
+          <Icon loading size="big" color={colorTwo} name={iconOuter} />
+        )}
+
+        <Popup
+          // className="popup-custom"
+          trigger={this.renderIcon(
+            userIcons,
+            data,
+            nodeId,
+            isOpen,
+            icon,
+            colorOne,
+            colorTwo
+          )}
+          on="click"
+          style={{ borderRadius: 0 }}
+          inverted
+          // onClose={() => this.setState({[`popup_${nodeId}`]:false})}
+          // onOpen={() => this.setState({[`popup_${nodeId}`]:true})}
+          open={this.state[`popup_${nodeId}`] === true && closeCharts === false}
+          content={this.renderChart(data.mainChart)}
+          position="bottom center"
+        />
+      </Icon.Group>
+    );
+  };
+
+  render() {
+    const { node, nodeSize } = this.props;
 
     return (
       <DataConsumer>
-        {({ userIcons }) => (
-          <div style={{ height: nodeSize / 10, width: nodeSize / 10 }}>
-            <div
-              className="centered"
-              style={{ height: nodeSize / 10, width: nodeSize / 10 }}
-            >
-              {renderIconGroup(
-                userIcons,
-                data,
-                node.id,
-                metrics,
-                icon,
-                colorTwo,
-                colorOne,
-                closeCharts
-              )}
+        {({ userIcons, mapData, closeCharts }) => {
+          const data = ((mapData || {}).nodeData || {})[node.id] || {};
+          const icon = data.icon || setEntityDesign(data.entityType).icon;
+          const metrics = buildNodeMetrics(data);
+
+          let { colorOne, colorTwo, iconOne } = setAlertDesign(
+            data.alertSeverity,
+            data.entityType
+          );
+          if (
+            data.customAlert &&
+            data.customAlertData &&
+            data.customAlertData[0]
+          ) {
+            const customAlertDesign = setCustomAlertDesign(
+              data.customAlert,
+              data.customAlertData
+            );
+            colorOne = customAlertDesign.colorOne;
+            colorTwo = customAlertDesign.colorTwo;
+          }
+
+          return (
+            <div style={{ height: nodeSize / 10, width: nodeSize / 10 }}>
+              <div
+                className="centered"
+                style={{ height: nodeSize / 10, width: nodeSize / 10 }}
+              >
+                {this.renderIconGroup(
+                  userIcons,
+                  data,
+                  node.id,
+                  metrics,
+                  icon,
+                  iconOne,
+                  colorTwo,
+                  colorOne,
+                  closeCharts
+                )}
+              </div>
             </div>
-          </div>
-        )}
+          );
+        }}
       </DataConsumer>
     );
   }
