+++ nerdlets/observability-maps-nerdlet/components/custom-nodes/handler.js
@@ -3,15 +3,7 @@ import DefaultNode from './default';
 
 export default class NodeHandler extends React.PureComponent {
   render() {
-    const { sidebarOpen, node, mapData, nodeSize, closeCharts } = this.props;
-    return (
-      <DefaultNode
-        sidebarOpen={sidebarOpen}
-        node={node}
-        mapData={mapData}
-        nodeSize={nodeSize}
-        closeCharts={closeCharts}
-      />
-    );
+    const { node, nodeSize } = this.props;
+    return <DefaultNode node={node} nodeSize={nodeSize} />;
   }
 }
